Move caller_module_and_locals() to IPython.util.frame
